feat: 예외처리 ViewModel로 처리
