+++ pennyway-client-iOS/pennyway-client-iOS/View/SignUpView/NavigationBackButton.swift
@@ -1,0 +1,19 @@
+
+
+import SwiftUI
+
+struct BackButton: View {
+    @Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
+    
+    var body: some View {
+        Button(action: {
+            self.presentationMode.wrappedValue.dismiss()
+        }, label: {
+            Image("icon_arrow_back")
+                .resizable()
+                .aspectRatio(contentMode: .fit)
+                .frame(width: 34, height: 34)
+                .padding(5)
+        })
+    }
+}
