--- pennyway-client-iOS/pennyway-client-iOS/ViewModel/viewmodel_ex_file.swift
@@ -1,8 +1,0 @@
-//
-//  ex_file.swift
-//  pennyway-client-iOS
-//
-//  Created by 최희진 on 3/7/24.
-//
-
-import Foundation
